Phase 5 — Video capture: record + play + audio/webcam + export (v0.60.0)
Screen recording to the owned freally-video (.fvid) format, with in-app
playback and optional export — across P5.1 (recording) and P5.2 (audio/webcam).

- freally-video: streaming encoder/decoder (4K/long recordings stay
  memory-bounded; byte-identical to the batch codec) + read_audio_file, plus
  decode-path DoS hardening (bound every allocation derived from file fields).
- freally-capture: efficient per-frame region grab + window-follow.
- app: recorder (worker-thread timed capture -> .fvid; pause/resume/stop;
  real-time frame duplication; window-follow; finalizes even on app exit), the
  Video button + capture-bar toggle, an always-on-top REC control bar, an
  in-app .fvid player (own decode path) opened from the recents strip, and
  Export to GIF (built-in) / WebM / MP4 (ffmpeg as a fetched-on-demand subprocess).
- audio (cpal): system loopback + microphone, mixed; webcam (nokhwa) PiP.
- Settings: record fps + audio/webcam toggles. DRAFT EULA + Privacy Policy.
- Docs: README/CHANGELOG/SECURITY/THIRD-PARTY-NOTICES + the Pages site; CI gains
  libv4l-dev. Version 0.45.0 -> 0.60.0.

Reviewed with /code-review + /security-review across P5.0/P5.1/P5.2; fixed the
ffmpeg-export deadlock, even-dimension export, off-screen Stop button,
lost-recording-on-exit, and the codec decode-path allocation DoS holes.

+++ crates/video/src/inter.rs
@@ -18,7 +18,7 @@ const TILE: u32 = 16;
 pub(crate) fn encode(prev: &[u8], cur: &[u8], width: u32, height: u32) -> (Vec<u8>, usize, usize) {
     let tiles_x = width.div_ceil(TILE);
     let tiles_y = height.div_ceil(TILE);
-    let total = (tiles_x * tiles_y) as usize;
+    let total = tiles_x as usize * tiles_y as usize;
 
     let mut dirty = vec![false; total];
     let mut dirty_count = 0;
@@ -72,7 +72,7 @@ pub(crate) fn decode(stream: &[u8], prev: &[u8], width: u32, height: u32) -> Opt
     }
     let tiles_x = width.div_ceil(tile);
     let tiles_y = height.div_ceil(tile);
-    let total = (tiles_x * tiles_y) as usize;
+    let total = tiles_x as usize * tiles_y as usize;
     let bitmap_len = total.div_ceil(8);
 
     let bitmap = stream.get(1..1 + bitmap_len)?;
